requirements.txt
asgiref==3.7.2
Django==4.2.9
django-cloudflare==0.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
